fixed the bell icon
