+++ package.json
@@ -26,6 +26,7 @@
     "react-loading": "^2.0.3",
     "react-router-dom": "^6.11.1",
     "react-scripts": "5.0.1",
+    "reactstrap": "^9.1.10",
     "styled-components": "^5.3.10",
     "web-vitals": "^2.1.4"
   },
