+++ src/components/Navigation/NavBar.css
@@ -8,3 +8,9 @@
 
 
   
+.bell-icon{
+    height: 30px;
+    width: 30px;
+    margin-top: 25px;
+    margin-left: 10px;
+}
