+++ src/components/Products/Import/ImportItem.js
@@ -1,43 +1,41 @@
-import React, { useState, useEffect } from "react";
-import { useNavigate, useParams } from "react-router-dom";
+import React from "react";
+import { useNavigate } from "react-router";
 import axios from "axios";
+import { useState, useEffect } from "react";
 import Button from "react-bootstrap/Button";
+import { useParams } from "react-router-dom";
 import { FaFacebook, FaInstagram, FaTwitter } from "react-icons/fa";
 import moment from "moment";
 import NavBar from "../../Navigation/NavBar";
-import LoadingBar from "../../LoadingScreens/LoadingBar"; 
 
 const ImportItem = () => {
   const navigate = useNavigate();
   const [importProduct, setImportProduct] = useState([]);
-  const [isLoading, setIsLoading] = useState(true); 
-  const { id } = useParams();
+  //eslint-disable-next-line
+  const [loading, setLoading] = useState(false);
+  const [tokens, setTokens] = useState(localStorage.getItem("tokens") || 0);
+  const chatprice = tokens - 10;
+  let userId = localStorage.getItem("userId");
 
   const getImportProduct = async () => {
-    try {
-      const data = {
-        headers: {
-          Accept: "application/json",
-        },
-      };
+    const data = {
+      headers: {
+        Accept: "application/json",
+      },
+    };
 
-      const response = await axios.get(
-        `http://127.0.0.1:8000/api/ilist/${id}`,
-        data
-      );
-      const apiImportProducts = response.data.data;
-      setImportProduct(apiImportProducts);
-      setIsLoading(false); // Set loading state to false when data is fetched
-    } catch (error) {
-      console.error("Error fetching import product:", error);
-      setIsLoading(false); // Set loading state to false on error
-    }
+    const response = await axios.get(
+      "http://127.0.0.1:8000/api/ilist/" + id,
+      data
+    );
+    const apiImportProducts = response.data.data;
+    setImportProduct(apiImportProducts);
   };
-
+  let { id } = useParams();
   useEffect(() => {
-    getImportProduct();
+    getImportProduct(id);
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  }, [id]);
 
   const handleBack = () => {
     navigate("/Import");
@@ -58,31 +56,53 @@ const ImportItem = () => {
     }
   };
 
-  if (isLoading) {
-    // Render loading screen while data is being fetched
-    return <LoadingBar />;
-  }
+  const handleChat = async (id) => {
+    if (tokens < 10) {
+      alert("You should have at least 10 tokens to chat with the owner!");
+    } else {
+      const confirmChat = window.confirm(
+        "Chat with the owner? This will take 10 tokens from your account."
+      );
+
+      if (confirmChat) {
+        setLoading(true);
+        try {
+          await axios.put(`http://localhost:8000/api/updateToken/${userId}`, 
+          {
+            amount: chatprice,
+          });
+          setLoading(false);
+          navigate("/ContactFrom/" + id);
+        } catch (error) {
+          setLoading(false);
+        }
+        const response = await axios.get(
+          `http://localhost:8000/api/token/${userId}`
+        );
+        setTokens(response.data.amount);
+      }
+    }
+  };
 
   return (
     <div>
       <NavBar />
-      <div className="d-flex justify-content-center  mt-4 text-primary">
+      <div className="d-flex justify-content-center mt-4 text-primary">
         <h1>Import Details</h1>
       </div>
       <div>
-        <div className="col-xl-5 col-lg-6 col-md-8 col-sm-10 mx-auto   mb-4 p-5 border rounded  border-dark ">
+        <div className="col-xl-5 col-lg-6 col-md-8 col-sm-10 mx-auto mb-4 p-5 border rounded border-dark">
           {importProduct.map((importProduct) => {
             return (
               <div key={importProduct.id}>
                 <div>
-                  <p>Name: {importProduct.name}</p>
-                  <p>Country: {importProduct.country}</p>
-                  <p>Price: {importProduct.price}</p>
-                  
-                  <p>Description: {importProduct.description}</p>
-                  <p>Added: {formatDate(importProduct.created_at)}</p>
-                  <p>Views: {importProduct.views}</p>
-                  <p>Category: {importProduct.category_name}</p>
+                  <p>country: {importProduct.country}</p>
+                  <p>price: {importProduct.price}</p>
+                  <p>name: {importProduct.name}</p>
+                  <p>description: {importProduct.description}</p>
+                  <p>created at: {formatDate(importProduct.created_at)}</p>
+                  <p>views: {importProduct.views}</p>
+                  <p>category: {importProduct.category_name}</p>
                   <a href="https://www.facebook.com/" className="m-2">
                     <FaFacebook />
                   </a>
@@ -94,7 +114,12 @@ const ImportItem = () => {
                   </a>
                 </div>
                 <div className="d-flex justify-content-center btn-lg">
-                  <Button onClick={handleBack}>Back</Button>
+                  <Button className="mx-3" onClick={handleBack}>
+                    Back
+                  </Button>
+                  <Button onClick={() => handleChat(importProduct.id)}>
+                    Chat with owner
+                  </Button>
                 </div>
               </div>
             );
